Added function to vary alpha randomly
Added a function that samples the value of the water abundance factor between 0.7 and 1, adjusts the cell volume and diameter and then plots a Lowess fitted curve showing cell diameter vs ploidy. The cell diameter is also plotted for alpha = 0.7 and alpha = 1.0. This is refactored from the separate R script that was used in the paper.